+++ src/services/pages/detail-service.ts
@@ -7,7 +7,7 @@ import { useParams } from "react-router";
 import { Observable } from "rxjs";
 import { finalize } from "rxjs/operators";
 import { formService } from "services/form-service";
-import { routerService } from "services/RouterService";
+import { routerService } from "services/route-service";
 import { v4 as uuidv4 } from "uuid";
 import appMessageService from "services/app-message-service";
